Se ha cambiado en el create de la tabla competencias_actividades, para que en delete, se haga cascade y se han creado las policies de competencias

+++ app/Http/Controllers/API/CompetenciaController.php
@@ -5,11 +5,25 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\CompetenciaResource;
 use App\Models\Competencia;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 
-class CompetenciaController extends Controller
+class CompetenciaController extends Controller implements HasMiddleware
 {
     public $modelclass = Competencia::class;
+
+    /**
+     * Get the middleware that should be assigned to the controller.
+     */
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show']),
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -25,6 +39,8 @@ class CompetenciaController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Competencia::class);
+
         $competencia = json_decode($request->getContent(), true);
 
         $ciclo = Competencia::create($competencia);
@@ -45,6 +61,8 @@ class CompetenciaController extends Controller
      */
     public function update(Request $request, Competencia $competencia)
     {
+        Gate::authorize('update', $competencia);
+
         $competenciaData = json_decode($request->getContent(), true);
         $competencia->update($competenciaData);
 
@@ -56,6 +74,8 @@ class CompetenciaController extends Controller
      */
     public function destroy(Competencia $competencia)
     {
+        Gate::authorize('delete', $competencia);
+
         try {
             $competencia->delete();
             return response()->json(null, 204);
